Adding Terminal Crop Region for comparing images - Adding terminalHelper for handling Terminal location. - Adding imageUtil.CropRegion for Crop Region's naming. - Adding visualmetricsWrapper.find_terminal_view for comparing. - Modifying inputLatencyAnimationDctGenerator for finding "start" by Terminal Region
Few weeks ago, @ShakoHo, @MikeLien and I have a discussion on comparing images for finding events, e.g. "start" and "end".
We found the "start" event of Input Latency cases by comparing the full screenshot (original image).
However, the IL cases will print the LOG on the Terminal when triggering the action.
We only want to compare the Terminal Region to find the "start" event for preciseness.
That's why I try to crop the Terminal Region from Original images.