Refactoring the runtime, frame-throughput generator, and checking the error of imageUtile
We found the "time_sequence" of run_time_dict is no used, so I try to store all event time base on the start time, and then rename to "event_time".

And then, sometimes CV2 will fail when convert the video to images.
It will fail the search range when images length less then search range start index.
e.g:
When search range are [3000, 4000, 3500, 4500] with image length 2001.
It will be fixed to [3000, 2000, 2500, 2000] in imageUtil.get_search_range() method.
We should check the image length before starting the comparing.