Support multiple-recorder for capturing 2nd snapshot at teardown
From PR #611, we implement `commonUtil.is_video_recording` for checking video recorders.
So that will supports both `avconvProfiler` (FFmpeg) and `obsProfiler` (OBS Studio).
However, we only change the condition at `setUp`, the condition at `tearDown` was not changed.
It will make the Hasal miss the 2nd screenshot when we enable `'snapshot-base-sample2'` in index-config with OBS.