Moving the running time calculation method from `commonUtil.CalculationUtil` to `basGenerator`
After having a discussion with @ShakoHo and @MikeLien, we think it will be more reasonable if we move the running time calculation method from CalculationUtil to baseGenerator.
When sub-generators have to customize their own calculating methods, then they can override the "calculate_runtime_base_on_event" to fit the requirements.